Fixed bugs, settings refactored
