+++ README.md
@@ -118,6 +118,7 @@ You can configure Chibi using the following environment variables:
 | OPENAI_TEMPERATURE           | OpenAI temperature for response generation                                                                                                        | No       | 0.5                                                                              |
 | PROXY                        | Proxy settings for your application                                                                                                               | No       |                                                                                  |
 | REDIS                        | Redis connection string, i.e. "redis://localhost"                                                                                                 | No       |                                                                                  |
+| REDIS_PASSWORD               | Redis password (optional)                                                                                                                         | No       |                                                                                  |
 | TIMEOUT                      | Timeout (in seconds) for processing requests                                                                                                      | No       | 15                                                                               |
 | USERS_WHITELIST              | Comma-separated list of whitelisted usernames, i.e. `"@YourName,@YourFriendName,@YourCatName"`                                                    | No       |                                                                                  |
