+++ chibi/config/__init__.py
@@ -1,0 +1,10 @@
+import chibi.config.logging  # noqa: F401
+from chibi.config.app import application_settings
+from chibi.config.gpt import gpt_settings
+from chibi.config.telegram import telegram_settings
+
+__all__ = [
+    'application_settings',
+    'gpt_settings',
+    'telegram_settings'
+]
