+++ examples/.example.env
@@ -12,6 +12,7 @@ TELEGRAM_BOT_TOKEN="XXX"
 #MAX_HISTORY_TOKENS=1500
 #MESSAGE_FOR_DISALLOWED_USERS="You're not allowed to interact with me, sorry. Contact my owner first, please."
 #REDIS="redis://localhost"
+#REDIS_PASSWORD="Secret!"
 #TIMEOUT=15
 
 ## OpenAI Client Settings
