+++ examples/docker-compose.personal-use.yml
@@ -1,0 +1,20 @@
+version: '3'
+
+services:
+  chibi:
+   restart: unless-stopped
+   image: pysergio/chibi:latest
+   environment:
+     OPENAI_API_KEY: "sk-XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
+     TELEGRAM_BOT_TOKEN: "9999999999:AAAAAAAAAAAAAAAAAAAA-BBBBBBBBBBBBBB"
+     USERS_WHITELIST: "@YouUserName,@YourFriendUserName"
+     REDIS: "redis://redis"
+
+  redis:
+    restart: unless-stopped
+    image: redis:6-alpine
+    volumes:
+      - redis-data:/data
+
+volumes:
+  redis-data:
