+++ examples/docker-compose.public-use.yml
@@ -1,0 +1,18 @@
+version: '3'
+
+services:
+  chibi:
+   restart: unless-stopped
+   image: pysergio/chibi:latest
+   environment:
+     TELEGRAM_BOT_TOKEN: "9999999999:AAAAAAAAAAAAAAAAAAAA-BBBBBBBBBBBBBB"
+     REDIS: "redis://redis"
+
+  redis:
+    restart: unless-stopped
+    image: redis:6-alpine
+    volumes:
+      - redis-data:/data
+
+volumes:
+  redis-data:
