+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "chibi"
-version = "0.4.0"
+version = "0.5.0"
 description = "Another one ChatGPT Telegram bot"
 authors = ["Sergei Nagaev <nagaev.sv@gmail.com>"]
 license = "MIT"
@@ -9,9 +9,9 @@ readme = "README.md"
 [tool.poetry.dependencies]
 python = "^3.9"
 aioredis = "2.0.1"
-loguru = "0.6.0"
+loguru = "0.7.0"
 openai = "0.27.4"
-pydantic = "1.10.7"
+pydantic = {extras = ["dotenv"], version = "1.10.7"}
 python-telegram-bot = "20.2"
 
 [tool.poetry.group.dev.dependencies]
