+++ requirements.txt
@@ -351,7 +351,7 @@ multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0" \
 openai==0.27.4 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:3b82c867d531e1fd2003d9de2131e1c4bfd4c70b1a3149e0543a555b30807b70 \
     --hash=sha256:9f9d27d26e62c6068f516c0729449954b5ef6994be1a6cbfe7dbefbc84423a04
-pydantic==1.10.7 ; python_version >= "3.9" and python_version < "4.0" \
+pydantic[dotenv]==1.10.7 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:01aea3a42c13f2602b7ecbbea484a98169fb568ebd9e247593ea05f01b884b2e \
     --hash=sha256:0cd181f1d0b1d00e2b705f1bf1ac7799a2d938cce3376b8007df62b29be3c2c6 \
     --hash=sha256:10a86d8c8db68086f1e30a530f7d5f83eb0685e632e411dbbcf2d5c0150e8dcd \
@@ -388,6 +388,9 @@ pydantic==1.10.7 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:e79e999e539872e903767c417c897e729e015872040e56b96e67968c3b918b2d \
     --hash=sha256:ecbbc51391248116c0a055899e6c3e7ffbb11fb5e2a4cd6f2d0b93272118a209 \
     --hash=sha256:f4a2b50e2b03d5776e7f21af73e2070e1b5c0d0df255a827e7c632962f8315af
+python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0" \
+    --hash=sha256:a8df96034aae6d2d50a4ebe8216326c61c3eb64836776504fcca410e5937a3ba \
+    --hash=sha256:f5971a9226b701070a4bf2c38c89e5a3f0d64de8debda981d1db98583009122a
 python-telegram-bot==20.2 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:4d1d4b643ce158aa17a0987b84005eaf25fe0ce8b38fd234099594985611c198 \
     --hash=sha256:d0aa53e1f06d7cb7919cc0e2d6c81a02d968fc29921aeaa962edd1efb816a9bd
